Move Writer into its internal module.
