+++ src/Control/Effect/Writer/Internal.hs
@@ -1,2 +1,27 @@
+{-# LANGUAGE DeriveFunctor #-}
+{-# LANGUAGE ExistentialQuantification #-}
+{-# LANGUAGE StandaloneDeriving #-}
 module Control.Effect.Writer.Internal
-() where
+( Writer(..)
+) where
+
+import Control.Effect.Class
+
+data Writer w m k
+  = Tell w (m k)
+  | forall a . Listen (m a) (w -> a -> m k)
+  | forall a . Censor (w -> w) (m a) (a -> m k)
+
+deriving instance Functor m => Functor (Writer w m)
+
+instance HFunctor (Writer w) where
+  hmap f (Tell w     k) = Tell w         (f       k)
+  hmap f (Listen   m k) = Listen   (f m) ((f .) . k)
+  hmap f (Censor g m k) = Censor g (f m) (f     . k)
+  {-# INLINE hmap #-}
+
+instance Effect (Writer w) where
+  handle state handler (Tell w     k) = Tell w                          (handler (k <$ state))
+  handle state handler (Listen   m k) = Listen   (handler (m <$ state)) (fmap handler . fmap . k)
+  handle state handler (Censor f m k) = Censor f (handler (m <$ state)) (handler . fmap k)
+  {-# INLINE handle #-}
